backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.30.0
databases[asyncpg]==0.9.0
sqlalchemy==2.0.41
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
google-generativeai==0.5.4
pandas==2.2.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.11.0
pydantic-settings==2.9.0
httpx==0.27.0
aiofiles==23.2.1
reportlab==4.1.0
pillow==10.4.0
